requirements.txt
Django==2.2.4
django-crispy-forms==1.7.2
django-filter==2.3.0
html5lib==1.0.1
Pillow==7.1.2
PyPDF2==1.26.0
pytz==2019.2
reportlab==3.5.42
six==1.12.0
sqlparse==0.3.0
webencodings==0.5.1
xhtml2pdf==0.2.4
